+++ lib/msal4flutter.dart
@@ -1,3 +1,5 @@
 export 'src/clients.dart';
 export 'src/device_code_response.dart';
+export 'src/token_cache.dart';
 export 'src/token_response.dart';
+export 'src/token_storage.dart';